modularizing the modals
